feature(core) add custom serializerInterface
