+++ module/api/src/Application/EventListener/RequestBodyListener.php
@@ -12,7 +12,7 @@ namespace Ergonode\Api\Application\EventListener;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
-use Ergonode\Core\Infrastructure\Serializer\SerializerInterface;
+use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 class RequestBodyListener
 {
