+++ module/attribute/src/Infrastructure/Persistence/Projector/Attribute/AbstractDbalAttributeParameterChangeEventProjector.php
@@ -12,7 +12,7 @@ namespace Ergonode\Attribute\Infrastructure\Persistence\Projector\Attribute;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
-use Ergonode\Core\Infrastructure\Serializer\SerializerInterface;
+use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 abstract class AbstractDbalAttributeParameterChangeEventProjector
 {
