+++ module/channel/src/Infrastructure/Persistence/Repository/Mapper/DbalChannelMapper.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
 namespace Ergonode\Channel\Infrastructure\Persistence\Repository\Mapper;
 
 use Ergonode\Channel\Domain\Entity\AbstractChannel;
-use Ergonode\Core\Infrastructure\Serializer\SerializerInterface;
+use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 class DbalChannelMapper
 {
