+++ module/designer/src/Infrastructure/Persistence/Projector/DbalTemplateElementChangedEventProjector.php
@@ -11,7 +11,7 @@ namespace Ergonode\Designer\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
 use Ergonode\Designer\Domain\Event\TemplateElementChangedEvent;
-use Ergonode\Core\Infrastructure\Serializer\SerializerInterface;
+use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 class DbalTemplateElementChangedEventProjector
 {
