+++ module/importer/src/Infrastructure/Persistence/Repository/Factory/DbalSourceFactory.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
 namespace Ergonode\Importer\Infrastructure\Persistence\Repository\Factory;
 
 use Ergonode\Importer\Domain\Entity\Source\AbstractSource;
-use Ergonode\Core\Infrastructure\Serializer\SerializerInterface;
+use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 class DbalSourceFactory
 {
