+++ module/product-collection/src/Infrastructure/Persistence/Projector/ProductCollection/DbalProductCollectionNameChangedEventProjector.php
@@ -13,7 +13,7 @@ use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
 use Doctrine\DBAL\Types\Types;
 use Ergonode\ProductCollection\Domain\Event\ProductCollectionNameChangedEvent;
-use Ergonode\Core\Infrastructure\Serializer\SerializerInterface;
+use Ergonode\Core\Application\Serializer\SerializerInterface;
 
 class DbalProductCollectionNameChangedEventProjector
 {
